:recycle: (useModalStyle) remove file
to move to styled component
