+++ src/ui/components/modals/AppModal.tsx
@@ -1,16 +1,15 @@
-import React, { FunctionComponent, useRef, useState } from 'react'
-import { ScrollView, useWindowDimensions, View } from 'react-native'
+import React, { FunctionComponent, useRef, useState, useMemo } from 'react'
+import { ScrollView, useWindowDimensions, View, StyleSheet } from 'react-native'
 import RNModal from 'react-native-modal'
 import styled from 'styled-components/native'
 
 import { useKeyboardEvents } from 'ui/components/keyboard/useKeyboardEvents'
 import { Style } from 'ui/components/Style'
 import { IconInterface } from 'ui/svg/icons/types'
-import { getSpacing, UniqueColors } from 'ui/theme'
+import { getSpacing, UniqueColors, ColorsEnum } from 'ui/theme'
 import { useCustomSafeInsets } from 'ui/theme/useCustomSafeInsets'
 
 import { ModalHeader } from './ModalHeader'
-import { ModalStyles, useModalStyles } from './useModalStyles'
 
 interface Props extends ModalStyles {
   title: string
@@ -27,6 +26,20 @@ interface Props extends ModalStyles {
 }
 
 const webcss = `div[aria-modal="true"] { align-items: center }`
+
+const defaultModalStyles = {
+  spacing: getSpacing(1),
+  height: getSpacing(112),
+  maxWidth: getSpacing(125),
+  layout: undefined,
+}
+
+export interface ModalStyles {
+  layout?: 'bottom'
+  height?: number
+  maxWidth?: number
+  spacing?: number
+}
 
 export const AppModal: FunctionComponent<Props> = ({
   height,
@@ -46,6 +59,63 @@ export const AppModal: FunctionComponent<Props> = ({
   onBackdropPress,
 }) => {
   const { height: windowHeight, width: windowWidth } = useWindowDimensions()
+
+  const useModalStyles = (modalStyles: ModalStyles) => {
+    const { layout, spacing, height, maxWidth } = { ...defaultModalStyles, ...modalStyles }
+    return useMemo(
+      () =>
+        StyleSheet.create({
+          topOffset: {
+            ...(layout === 'bottom'
+              ? {
+                  position: 'absolute',
+                  height,
+                  margin: 'auto',
+                  bottom: 0,
+                }
+              : {}),
+          },
+          topInset: {
+            ...(layout === 'bottom'
+              ? {
+                  marginTop: 0,
+                }
+              : {}),
+          },
+          container: {
+            maxWidth,
+            ...(layout === 'bottom'
+              ? {
+                  marginBottom: 0,
+                  marginRight: 0,
+                  marginLeft: 0,
+                  borderTopRightRadius: 20,
+                  borderTopLeftRadius: 20,
+                }
+              : {
+                  marginHorizontal: 'auto',
+                  borderTopRightRadius: 20,
+                  borderTopLeftRadius: 20,
+                  borderBottomLeftRadius: 20,
+                  borderBottomRightRadius: 20,
+                }),
+            flexDirection: 'column',
+            backgroundColor: ColorsEnum.WHITE,
+            justifyContent: 'center',
+            alignItems: 'center',
+            width: '100%',
+            borderTopStartRadius: getSpacing(4),
+            borderTopEndRadius: getSpacing(4),
+            padding: getSpacing(6),
+          },
+          contentContainer: {},
+          modaleIcon: {
+            marginTop: 20,
+          },
+        }),
+      [layout, spacing, height, maxWidth]
+    )
+  }
   const styles = useModalStyles({
     layout: layout || 'bottom',
     height,
@@ -81,6 +151,7 @@ export const AppModal: FunctionComponent<Props> = ({
         backdropColor={UniqueColors.GREY_OVERLAY}
         isVisible={visible}
         onBackdropPress={handleOnBackdropPress()}
+        // TODO : remove stylesheet and use styledcomponents
         style={[styles.container, styles.topOffset]}
         testID="modal"
         deviceHeight={windowHeight}
